Crate relay_redis

Abstraction over Redis caches.

By default, this library only implements an empty noop client. With the impl feature, the actual Redis client is implemented.

Re-exports

pub use redis;

Structs

Connection

A reference to a pooled connection from PooledClient.

RedisClusterConnectionManager

[bb8::ManageConnection] for an async redis cluster.

RedisConfigOptions

Additional configuration options for a redis client.

RedisPools

The various RedisPools used within Relay.

RedisScripts

A collection of static methods to load predefined Redis scripts.

ResponseTimeoutCustomConnection

[bb8::CustomizeConnection] that calls set_response_timeout on connections.

Stats

Stats about how the RedisPool is performing.

Enums

AsyncRedisPool

An abstraction over Single and Cluster redis connections.

PooledClient

A pooled Redis client.

RedisError

An error returned from RedisPool.

RedisPool

Abstraction over cluster vs non-cluster mode.