Crate relay_redis

Abstraction over Redis caches.

By default, this library only implements an empty noop client. With the impl feature, the actual Redis client is implemented.

Re-exports

pub use redis;

Structs

AsyncRedisClient

Asynchronous redis client that wraps an AsyncRedisConnection.

RedisClientStats

Stats about how the Redis client is performing.

RedisClients

The various AsyncRedisClients used by Relay.

RedisConfigOptions

Additional configuration options for a redis client.

RedisScripts

A collection of static methods to load predefined Redis scripts.

Enums

AsyncRedisConnection

A wrapper type for async redis connections.

RedisError

An error returned from RedisPool.