Crate relay_redis

Abstraction over Redis caches.

By default, this library only implements an empty noop client. With the impl feature, the actual Redis client is implemented.

Re-exports

pub use redis;

Structs

AsyncRedisClient

Client that wraps a AsyncRedisConnection.

Connection

A reference to a pooled connection from PooledClient.

RedisConfigOptions

Additional configuration options for a redis client.

RedisPools

The various RedisPools used within Relay.

RedisScripts

A collection of static methods to load predefined Redis scripts.

Stats

Stats about how the RedisPool is performing.

Enums

AsyncRedisConnection

A wrapper Type for async redis connections. Conceptually it’s similar to RedisPool but async redis does not require a pool since the connections can just be cloned and are thread safe.

PooledClient

A pooled Redis client.

RedisError

An error returned from RedisPool.

RedisPool

Abstraction over cluster vs non-cluster mode.