Crate relay_redis

Abstraction over Redis caches.

By default, this library only implements an empty noop client. With the impl feature, the actual Redis client is implemented.

Re-exports

pub use redis;

Structs

RedisClientStats

Statistics about the Redis client’s connection client state.

RedisClients

A collection of Redis clients used by Relay for different purposes.

RedisConfigOptions

Additional configuration options for a redis client.

RedisScripts

A collection of static methods to load predefined Redis scripts.

Enums

AsyncRedisClient

A connection client that can manage either a single Redis instance or a Redis cluster.

AsyncRedisConnection

A connection to either a single Redis instance or a Redis cluster.

RedisError

An error type that represents various failure modes when interacting with Redis.

Type Aliases

Result

Typical Result when dealing with Redis.